pyproject.toml
[tool.poetry]
name = "nsfw-classifier"
version = "0.1.0"
description = ""
authors = ["jy-choi <jy.choi@toss.im>"]
readme = "README.md"
packages = [{include = "nsfw_classifier"}]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.6.0"
torchvision = "^0.21.0"
pandas = "^2.2.3"
pyyaml = "^6.0.2"
pillow = "^11.1.0"
omegaconf = "^2.3.0"
tqdm = "^4.67.1"
onnx = "^1.17.0"
onnxruntime = "^1.20.1"
fastapi = "^0.115.8"
python-multipart = "^0.0.20"
uvicorn = "^0.34.0"
locust = "^2.32.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
